requirements.txt
Flask
dotenv
supabase
baidu-aip